requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astunparse==1.6.3
backtrader==1.9.75.123
beautifulsoup4==4.9.0
blis==0.4.1
Brotli==1.0.7
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
dash==1.14.0
dash-bootstrap-components==0.10.3
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
dataclasses==0.6
distlib==0.3.1
fastcore==1.0.0
fastprogress==1.0.0
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
fredapi==0.4.2
future==0.18.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mplfinance==0.12.6a3
multitasking==0.0.9
murmurhash==1.0.2
nodejs==0.1.1
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.3.0
optional-django==0.1.0
packaging==20.4
pandas==1.0.5
pandas-datareader==0.8.1
patsy==0.5.1
Pillow==7.2.0
plac==1.1.3
plotly==4.9.0
preshed==3.0.2
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.8.18
pydantic==1.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scipy==1.4.1
selenium==3.141.0
six==1.14.0
soupsieve==2.0
spacy==2.3.2
srsly==1.0.2
statsmodels==0.11.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
thinc==7.4.1
tqdm==4.48.0
urllib3==1.25.9
virtualenv==20.0.30
wasabi==0.7.1
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
yahoo-earnings-calendar==0.6.0
yahoofinancials==1.5
yahooquery==2.2.5
yfinance==0.1.54
